add tube_1d geom
