+++ tube/systems/tube/__init__.py
@@ -1,18 +1,22 @@
 # Copyright (c) 2023, twiinIT - All Rights Reserved
 # twiinIT proprietary - See licence file packaged with this code
 
+from tube.systems.tube.tube_1d_aero import Tube1DAero
+from tube.systems.tube.tube_1d_geom import Tube1DGeom
 from tube.systems.tube.tube_aero import TubeAero
 from tube.systems.tube.tube_geom import TubeGeom
 from tube.systems.tube.tube_mech import TubeMech
 
-from tube.systems.tube.tube_1d_aero import Tube1DAero  # isort: skip
 from tube.systems.tube.tube import Tube  # isort: skip
+from tube.systems.tube.tube_1d import Tube1D  # isort: skip
 
 
 __all__ = [
     "TubeAero",
     "Tube1DAero",
     "TubeGeom",
+    "Tube1DGeom",
     "TubeMech",
     "Tube",
+    "Tube1D",
 ]
